ingestion/requirements.txt
requests
azure-storage-blob
python-dotenv